notebook: remove use of Promises
Promises were added thinking the Stream would eventually use them.
But as a design choice, it is simpler to use the model of a memory cache of objects per stream.
A special version of Stream could be used if the server stores lots of data in a stream.

This change requires updating example notebook data if any previously loaded.